requirements.txt
Flask
pymongo
pandas
matplotlib
seaborn
python-dotenv
werkzeug
